pyproject.toml
[build-system]
requires = [
    "setuptools>=45.0",
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/fmu_settings_api/_version.py"

[project]
name = "fmu-settings-api"
description = "The API for fmu-settings"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0" }
authors = [
    { name = "Equinor", email = "fg-fmu_atlas@equinor.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering",
    "Topic :: Utilities",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Natural Language :: English",
]
dynamic = ["version"]
dependencies = [
    "fastapi",
    "fmu-datamodels>=0.21.2",  # unmappable relation, no equivalent relation
    "fmu-settings>=0.29.0",  # current Mappings schema
    "httpx",
    "packaging",
    "pydantic",
    "pyyaml",
    "rapidfuzz>=3.0.0",
    "runrms>=0.7.3",
    "structlog",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "python-dotenv",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xdist",
    "ruff",
    "types-PyYAML",
    "types-requests",
    "watchfiles",
    "fmu-settings-cli",
]

[project.urls]
Homepage = "https://github.com/equinor/fmu-settings-api"
Repository = "https://github.com/equinor/fmu-settings-api"
Documentation = "https://github.com/equinor/fmu-settings-api"

[tool.coverage.run]
omit = [
    "_version.py",
]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_subclassing_any = true
extra_checks = true
strict_equality = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
norecursedirs = [
    ".git",
    ".tox",
    ".env",
    "dist",
    "build",
]
markers = [
    "integration: Marks a test as an integration test",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
exclude = ["_version.py"]

[tool.ruff.lint]
ignore = [
    "C901",
]
select = [
    "B",  # flake-8-bugbear
    "C90",  # mccabe
    "D",  # pydocstyle
    "E",  # pycodestyle-error
    "F",  # pyflakes
    "I",  # isort
    "NPY",  # numpy
    "PD",  # pandas-vet
    "PIE",  # flake8-pie
    "Q",  # flake-8 quotes
    "Q",  # flake8-quotes
    "RET",  # flake8-return
    "RSE",  # flake8-raise
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "TID",  # flake8-tidy-imports
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-third-party = ["fmu.settings"]

[tool.ruff.lint.pydocstyle]
convention = "google"
